[BACKEND] - Criação do endpoint para Atualizar os dados do Tecnico e alteração nos valores de Long para Integer.

+++ src/main/java/com/valdir/helpdesk/domain/Cliente.java
@@ -21,7 +21,7 @@ public class Cliente extends Pessoa {
         addPerfil(Perfil.CLIENTE);
     }
 
-    public Cliente(Long id, String nome, String cpf, String email, String senha) {
+    public Cliente(Integer id, String nome, String cpf, String email, String senha) {
         super(id, nome, cpf, email, senha);
         addPerfil(Perfil.CLIENTE);
     }
